+++ package.json
@@ -5,7 +5,7 @@
   "dependencies": {
     "colors": "^1.4.0",
     "express": "^4.17.1",
-    "pm2": "^4.2.1",
+    "posix": "^4.1.2",
     "ws": "^7.2.0"
   },
   "devDependencies": {},
