+++ src/init.js
@@ -10,18 +10,24 @@
 
     module.exports = init;
 
+    /**
+     * Raises nofile limit on Linux system, and align config to arguments
+     */
     function init() {
         raise_nofile_limit();
         read_args();
     }
 
+    /**
+     * Raises Linux nofile limit
+     */
     function raise_nofile_limit() {
         let limit = null;
 
         try {
 
-            // 提升Linux系统nofile上限 (连接数量)
-            // 系统层面调整上限(Linux): /etc/security/limits.conf
+            // Raises Linux nofile limit
+            // System level configuration: /etc/security/limits.conf
             if (process.platform === 'linux') {
                 const posix = require('posix');
                 const hard_limit = posix.getrlimit('nofile')['hard'];
@@ -35,6 +41,10 @@
         return limit;
     };
 
+
+    /**
+     * Reads in arguments, overwrite configurations
+     */
     function read_args() {
         if (process.argv.includes('-v')) {
             config.verbose = true;
@@ -46,7 +56,7 @@
         config.ip = settings['wsServer']['self']['ip'];
         config.port = settings['wsServer']['self']['port'];
 
-        const ipIndex = process.argv.indexOf('--ip');
+        const ipIndex = process.argv.indexOf('--ws-ip');
         if (ipIndex !== -1) {
             const i = ipIndex;
             if (i + 1 < process.argv[i + 1]) {
@@ -55,7 +65,7 @@
             }
         }
 
-        const portIndex = process.argv.indexOf('--port');
+        const portIndex = process.argv.indexOf('--ws-port');
         if (portIndex !== -1) {
             const i = portIndex;
             if (i + 1 < process.argv[i + 1]) {
@@ -66,7 +76,7 @@
             }
         }
 
-        const httpIpIndex = process.argv.indexOf('--hIp');
+        const httpIpIndex = process.argv.indexOf('--http-ip');
         if (httpIpIndex !== -1) {
             const i = httpIpIndex;
             if (i + 1 < process.argv[i + 1]) {
@@ -75,7 +85,7 @@
             }
         }
 
-        const httpPortIndex = process.argv.indexOf('--hPort');
+        const httpPortIndex = process.argv.indexOf('--http-port');
         if (httpPortIndex !== -1) {
             const i = httpPortIndex;
             if (i + 1 < process.argv[i + 1]) {
